+++ packages/plugins/thorchain/src/tcPlugin.ts
@@ -13,6 +13,7 @@ import {
   TCBscDepositABI,
   TCEthereumVaultAbi,
   type UTXOChain,
+  type WalletChain,
   getMemoForLoan,
 } from "@swapkit/helpers";
 
@@ -143,7 +144,7 @@ function plugin({ getWallet, stagenet = false }: SwapKitPluginParams) {
         getMemoForLoan(type === "open" ? MemoType.OPEN_LOAN : MemoType.CLOSE_LOAN, {
           asset: assetValue.toString(),
           minAmount: minAmount.toString(),
-          address: getWallet(assetValue.chain).address,
+          address: getWallet(assetValue.chain as WalletChain).address,
         }),
     });
   }