+++ packages/swapkit/helpers/src/types/commonTypes.ts
@@ -1,5 +1,5 @@
 import type { RadixNetwork } from "@swapkit/toolbox-radix";
-import type { Chain } from "./chains";
+import type { Chain, WalletChain } from "./chains";
 import type { ChainApis } from "./sdk";
 import type { ChainWallet } from "./wallet";
 
@@ -78,10 +78,15 @@ export type ConnectConfig = SwapkitConfig & {
     applicationVersion: string;
     network: RadixNetwork;
   };
+
+  /**
+   * @optional for setting the kado api key
+   */
+  kadoApiKey?: string;
 };
 
 export type ConnectWalletParams<M = { [key in string]: any }> = {
-  addChain: <T extends Chain>(params: ChainWallet<T> & M) => void;
+  addChain: <T extends WalletChain>(params: ChainWallet<T> & M) => void;
   apis: ChainApis;
   config: ConnectConfig;
   rpcUrls: { [chain in Chain]?: string };