+++ packages/swapkit/helpers/src/types/derivationPath.ts
@@ -55,7 +55,8 @@ export const NetworkDerivationPath: Record<Chain, DerivationPathArray> = {
   THOR: [44, 931, 0, 0, 0],
 
   // Polkadot and related network derivation path is not number based
-  XRD: [0, 0, 0, 0, 0],
   DOT: [0, 0, 0, 0, 0],
+  FIAT: [0, 0, 0, 0, 0],
   FLIP: [0, 0, 0, 0, 0],
+  XRD: [0, 0, 0, 0, 0],
 };