+++ packages/swapkit/helpers/src/types/quotes.ts
@@ -159,6 +159,7 @@ export enum ProviderName {
   TRADERJOE_V2 = "TRADERJOE_V2",
   UNISWAP_V2 = "UNISWAP_V2",
   UNISWAP_V3 = "UNISWAP_V3",
+  KADO = "KADO",
 }
 
 export enum FeeTypeEnum {