+++ packages/swapkit/tokens/src/index.ts
@@ -1,6 +1,7 @@
 export { list as CaviarV1List } from "./tokenLists/caviar_v1";
 export { list as ChainflipList } from "./tokenLists/chainflip";
 export { list as JupiterList } from "./tokenLists/jupiter";
+export { list as KadoList } from "./tokenLists/kado";
 export { list as MayaList } from "./tokenLists/mayachain";
 // export { list as OciswapV1List } from "./tokenLists/ociswap_v1";
 export { list as OneInchList } from "./tokenLists/oneinch";