+++ packages/plugins/chainflip/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @swapkit/plugin-chainflip
 
+## 1.4.4
+
+### Patch Changes
+
+- Updated dependencies [[`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40)]:
+  - @swapkit/helpers@1.19.0
+  - @swapkit/api@2.0.1
+  - @swapkit/toolbox-evm@1.7.7
+  - @swapkit/toolbox-substrate@1.3.7
+  - @swapkit/toolbox-utxo@1.2.7
+
 ## 1.4.3
 
 ### Patch Changes