+++ packages/plugins/thorchain/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @swapkit/plugin-thorchain
 
+## 2.1.0
+
+### Minor Changes
+
+- [#1107](https://github.com/thorswap/SwapKit/pull/1107) [`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40) Thanks [@towanTG](https://github.com/towanTG)! - Adds Kado plugin
+
+### Patch Changes
+
+- Updated dependencies [[`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40)]:
+  - @swapkit/helpers@1.19.0
+  - @swapkit/api@2.0.1
+  - @swapkit/toolbox-cosmos@1.6.1
+  - @swapkit/toolbox-evm@1.7.7
+  - @swapkit/toolbox-substrate@1.3.7
+  - @swapkit/toolbox-utxo@1.2.7
+
 ## 2.0.0
 
 ### Major Changes