+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swapkit/api
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40)]:
+  - @swapkit/helpers@1.19.0
+
 ## 2.0.0
 
 ### Major Changes