+++ packages/swapkit/sdk/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @swapkit/sdk
 
+## 2.1.0
+
+### Minor Changes
+
+- [#1107](https://github.com/thorswap/SwapKit/pull/1107) [`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40) Thanks [@towanTG](https://github.com/towanTG)! - Adds Kado plugin
+
+### Patch Changes
+
+- Updated dependencies [[`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40)]:
+  - @swapkit/plugin-thorchain@2.1.0
+  - @swapkit/tokens@1.11.0
+  - @swapkit/plugin-radix@1.3.0
+  - @swapkit/plugin-kado@0.1.0
+  - @swapkit/core@3.1.0
+  - @swapkit/plugin-chainflip@1.4.4
+  - @swapkit/plugin-evm@1.3.4
+  - @swapkit/api@2.0.1
+  - @swapkit/wallets@1.2.38
+
 ## 2.0.12
 
 ### Patch Changes