+++ packages/wallets/polkadotjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @swapkit/wallet-polkadotjs
 
+## 1.0.49
+
+### Patch Changes
+
+- Updated dependencies [[`d88fe02`](https://github.com/thorswap/SwapKit/commit/d88fe02d8683a74a9b2f521b60381032099e5c40)]:
+  - @swapkit/helpers@1.19.0
+  - @swapkit/toolbox-substrate@1.3.7
+
 ## 1.0.48
 
 ### Patch Changes